+++ src/main/java/com/upokecenter/cbor/CBORUtilities.java
@@ -1443,15 +1443,15 @@ ValueNormalDays :
       floatExponent -= 1075;
       byte[] bytes = new byte[9];
       EInteger bigmantissa;
-      bytes[0] = (byte) (byte)(value0 & 0xff);
-      bytes[1] = (byte) (byte)((value0 >> 8) & 0xff);
-      bytes[2] = (byte) (byte)((value0 >> 16) & 0xff);
-      bytes[3] = (byte) (byte)((value0 >> 24) & 0xff);
-      bytes[4] = (byte) (byte)(value1 & 0xff);
-      bytes[5] = (byte) (byte)((value1 >> 8) & 0xff);
-      bytes[6] = (byte) (byte)((value1 >> 16) & 0xff);
-      bytes[7] = (byte) (byte)((value1 >> 24) & 0xff);
-      bytes[8] = (byte)0;
+      bytes[0] = (byte)(value0 & 0xff);
+      bytes[1] = (byte)((value0 >> 8) & 0xff);
+      bytes[2] = (byte)((value0 >> 16) & 0xff);
+      bytes[3] = (byte)((value0 >> 24) & 0xff);
+      bytes[4] = (byte)(value1 & 0xff);
+      bytes[5] = (byte)((value1 >> 8) & 0xff);
+      bytes[6] = (byte)((value1 >> 16) & 0xff);
+      bytes[7] = (byte)((value1 >> 24) & 0xff);
+      bytes[8] = 0;
       bigmantissa = EInteger.FromBytes(bytes, true);
       if (floatExponent == 0) {
         if (neg) {
